Rename: 유저 컴포넌트 폴더명 소문자 -> 대문자 수정
